# [Merged by Bors] - chore(algebra/direct_sum_graded): golf proofs