--- a/src/algebra/direct_sum_graded.lean
+++ b/src/algebra/direct_sum_graded.lean
@@ -318,7 +318,201 @@ begin
   exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_one ⟨i, xi⟩),
 end
 
+set_option profiler true
+
 private lemma mul_assoc (a b c : ⨁ i, A i) : a * b * c = a * (b * c) :=
+congr_at c $ congr_at b $ congr_at a $
+show (mul_hom A).comp_hom.comp (mul_hom A)               -- `λ a b c, a * b * c` as a bundled hom
+   = (add_monoid_hom.comp_hom flip_hom $
+       (mul_hom A).flip.comp_hom.comp (mul_hom A)).flip, -- `λ a b c, a * (b * c)` as a bundled hom
+begin
+  /-
+  ι : Type u_1,
+  _inst_1 : decidable_eq ι,
+  A : ι → Type u_2,
+  _inst_2 : Π (i : ι), add_comm_monoid (A i),
+  _inst_3 : add_monoid ι,
+  _inst_4 : gmonoid A,
+  a b c : ⨁ (i : ι), A i
+  ⊢ (⇑add_monoid_hom.comp_hom (mul_hom A)).comp (mul_hom A) =
+      (⇑(⇑add_monoid_hom.comp_hom flip_hom) ((⇑add_monoid_hom.comp_hom (mul_hom A).flip).comp (mul_hom A))).flip
+  -/
+
+  -- apply add_hom_ext, intros ai ax, apply add_hom_ext, intros bi bx, apply add_hom_ext, intros ci cx,
+  /-
+  ι : Type u_1,
+  _inst_1 : decidable_eq ι,
+  A : ι → Type u_2,
+  _inst_2 : Π (i : ι), add_comm_monoid (A i),
+  _inst_3 : add_monoid ι,
+  _inst_4 : gmonoid A,
+  a b c : ⨁ (i : ι), A i,
+  ai : ι,
+  ax : A ai,
+  bi : ι,
+  bx : A bi,
+  ci : ι,
+  cx : A ci
+  ⊢ ⇑(⇑(⇑((⇑add_monoid_hom.comp_hom (mul_hom A)).comp (mul_hom A)) (⇑(of (λ (i : ι), A i) ai) ax))
+           (⇑(of (λ (i : ι), A i) bi) bx))
+        (⇑(of (λ (i : ι), A i) ci) cx) =
+      ⇑(⇑(⇑((⇑(⇑add_monoid_hom.comp_hom flip_hom)
+                  ((⇑add_monoid_hom.comp_hom (mul_hom A).flip).comp (mul_hom A))).flip)
+              (⇑(of (λ (i : ι), A i) ai) ax))
+           (⇑(of (λ (i : ι), A i) bi) bx))
+        (⇑(of (λ (i : ι), A i) ci) cx)
+  -/
+
+  ext? ai ax bi bx ci cx : 6,
+  -- simp only [coe_comp, function.comp_app],
+  rw [coe_comp, function.comp_app, coe_comp, function.comp_app, coe_comp, function.comp_app,
+      coe_comp, function.comp_app, coe_comp, function.comp_app, coe_comp, function.comp_app,
+      coe_comp, function.comp_app],
+  /-
+  ι : Type u_1,
+  _inst_1 : decidable_eq ι,
+  A : ι → Type u_2,
+  _inst_2 : Π (i : ι), add_comm_monoid (A i),
+  _inst_3 : add_monoid ι,
+  _inst_4 : gmonoid A,
+  a b c : ⨁ (i : ι), A i,
+  ai : ι,
+  ax : A ai,
+  bi : ι,
+  bx : A bi,
+  ci : ι,
+  cx : A ci
+  ⊢ ⇑(⇑(⇑(⇑add_monoid_hom.comp_hom (mul_hom A)) (⇑(mul_hom A) (⇑(of (λ (i : ι), A i) ai) ax)))
+           (⇑(of (λ (i : ι), A i) bi) bx))
+        (⇑(of (λ (i : ι), A i) ci) cx) =
+      ⇑(⇑(⇑((⇑(⇑add_monoid_hom.comp_hom flip_hom)
+                  ((⇑add_monoid_hom.comp_hom (mul_hom A).flip).comp (mul_hom A))).flip)
+              (⇑(of (λ (i : ι), A i) ai) ax))
+           (⇑(of (λ (i : ι), A i) bi) bx))
+        (⇑(of (λ (i : ι), A i) ci) cx)
+  -/
+
+
+
+  -- ext ai ax bi bx ci cx : 6,
+  -- rw [coe_comp, function.comp_app, coe_comp, function.comp_app, coe_comp, function.comp_app,
+  --     coe_comp, function.comp_app, coe_comp, function.comp_app, coe_comp, function.comp_app],
+  /-
+  ι : Type u_1,
+  _inst_1 : decidable_eq ι,
+  A : ι → Type u_2,
+  _inst_2 : Π (i : ι), add_comm_monoid (A i),
+  _inst_3 : add_monoid ι,
+  _inst_4 : gmonoid A,
+  a b c : ⨁ (i : ι), A i,
+  ai : ι,
+  ax : A ai,
+  bi : ι,
+  bx : A bi,
+  ci : ι,
+  cx : A ci
+  ⊢ ⇑(⇑(⇑(⇑add_monoid_hom.comp_hom (mul_hom A)) (⇑(mul_hom A) (⇑(of (λ (i : ι), A i) ai) ax)))
+           (⇑(of (λ (i : ι), A i) bi) bx))
+        (⇑(of (λ (i : ι), A i) ci) cx) =
+      ⇑(⇑(⇑((⇑(⇑add_monoid_hom.comp_hom flip_hom)
+                  ((⇑add_monoid_hom.comp_hom (mul_hom A).flip).comp (mul_hom A))).flip.comp
+                 (of (λ (i : ι), A i) ai))
+              ax)
+           (⇑(of (λ (i : ι), A i) bi) bx))
+        (⇑(of (λ (i : ι), A i) ci) cx)
+  -/
+
+
+
+  -- ext ai ax bi bx ci cx : 6,
+  /-
+  ι : Type u_1,
+  _inst_1 : decidable_eq ι,
+  A : ι → Type u_2,
+  _inst_2 : Π (i : ι), add_comm_monoid (A i),
+  _inst_3 : add_monoid ι,
+  _inst_4 : gmonoid A,
+  a b c : ⨁ (i : ι), A i,
+  ai : ι,
+  ax : A ai,
+  bi : ι,
+  bx : A bi,
+  ci : ι,
+  cx : A ci
+  ⊢ ⇑((⇑((⇑(((⇑add_monoid_hom.comp_hom (mul_hom A)).comp (mul_hom A)).comp (of (λ (i : ι), A i) ai)) ax).comp
+               (of (λ (i : ι), A i) bi))
+            bx).comp
+           (of (λ (i : ι), A i) ci))
+        cx =
+      ⇑((⇑((⇑((⇑(⇑add_monoid_hom.comp_hom flip_hom)
+                    ((⇑add_monoid_hom.comp_hom (mul_hom A).flip).comp (mul_hom A))).flip.comp
+                   (of (λ (i : ι), A i) ai))
+                ax).comp
+               (of (λ (i : ι), A i) bi))
+            bx).comp
+           (of (λ (i : ι), A i) ci))
+        cx
+  -/
+
+
+
+  -- ext ai ax bi bx ci cx,
+  /-
+  ι : Type u_1,
+  _inst_1 : decidable_eq ι,
+  A : ι → Type u_2,
+  _inst_2 : Π (i : ι), add_comm_monoid (A i),
+  _inst_3 : add_monoid ι,
+  _inst_4 : gmonoid A,
+  a b c : ⨁ (i : ι), A i,
+  ai : ι,
+  ax : A ai,
+  bi : ι,
+  bx : A bi,
+  ci : ι,
+  cx : A ci,
+  i : ι
+  ⊢ ⇑(⇑((⇑((⇑(((⇑add_monoid_hom.comp_hom (mul_hom A)).comp (mul_hom A)).comp (of (λ (i : ι), A i) ai))
+                   ax).comp
+                  (of (λ (i : ι), A i) bi))
+               bx).comp
+              (of (λ (i : ι), A i) ci))
+           cx)
+        i =
+      ⇑(⇑((⇑((⇑((⇑(⇑add_monoid_hom.comp_hom flip_hom)
+                       ((⇑add_monoid_hom.comp_hom (mul_hom A).flip).comp (mul_hom A))).flip.comp
+                      (of (λ (i : ι), A i) ai))
+                   ax).comp
+                  (of (λ (i : ι), A i) bi))
+               bx).comp
+              (of (λ (i : ι), A i) ci))
+           cx)
+        i
+  -/
+
+
+  -- simplify lhs
+  -- rw [coe_comp, function.comp_app, comp_hom_apply_apply, coe_comp, function.comp_app],
+  rw [comp_hom_apply_apply, coe_comp, function.comp_app],
+  -- simplify rhs
+  rw [flip_apply, comp_hom_apply_apply, coe_comp, function.comp_app, flip_hom_apply, coe_comp,
+      function.comp_app, comp_hom_apply_apply, flip_apply, coe_comp, function.comp_app, flip_apply],
+  rw [of_mul_of', of_mul_of', of_mul_of', of_mul_of'],
+  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
+end
+
+private lemma mul_assoc' (a b c : ⨁ i, A i) : a * b * c = a * (b * c) :=
+congr_at c $ congr_at b $ congr_at a $
+show (mul_hom A).comp_hom.comp (mul_hom A)               -- `λ a b c, a * b * c` as a bundled hom
+   = (add_monoid_hom.comp_hom flip_hom $
+       (mul_hom A).flip.comp_hom.comp (mul_hom A)).flip, -- `λ a b c, a * (b * c)` as a bundled hom
+begin
+  ext ai ax bi bx ci cx : 6,
+  simp [of_mul_of'],
+  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
+end
+
+private lemma mul_assoc'' (a b c : ⨁ i, A i) : a * b * c = a * (b * c) :=
 congr_at c $ congr_at b $ congr_at a $
 show (mul_hom A).comp_hom.comp (mul_hom A)               -- `λ a b c, a * b * c` as a bundled hom
    = (add_monoid_hom.comp_hom flip_hom $
@@ -333,6 +527,8 @@ begin
   rw [of_mul_of', of_mul_of', of_mul_of', of_mul_of'],
   exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
 end
+
+
 
 /-- The `semiring` structure derived from `gmonoid A`. -/
 instance semiring : semiring (⨁ i, A i) := {