--- a/src/algebra/direct_sum_graded.lean
+++ b/src/algebra/direct_sum_graded.lean
@@ -300,8 +300,30 @@ variables [Π i, add_comm_monoid (A i)] [add_monoid ι] [gmonoid A]
 
 open add_monoid_hom (flip_hom coe_comp comp_hom_apply_apply flip_apply flip_hom_apply)
 
+-- #check flip add_monoid_hom.congr_fun
+
 private lemma one_mul (x : ⨁ i, A i) : 1 * x = x :=
 congr_at x $ show mul_hom A 1 = add_monoid_hom.id (⨁ i, A i),
+begin
+  apply add_hom_ext, intros i xi,
+  unfold has_one.one,
+  rw of_mul_of',
+  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.one_mul ⟨i, xi⟩),
+end
+
+private lemma one_mul' (x : ⨁ i, A i) : 1 * x = x :=
+begin
+  -- refine add_monoid_hom.congr_fun _ x,
+  refine add_monoid_hom.congr_fun (_ : mul_hom A 1 = add_monoid_hom.id (⨁ i, A i)) x,
+  apply add_hom_ext, intros i xi,
+  unfold has_one.one,
+  rw of_mul_of',
+  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.one_mul ⟨i, xi⟩),
+end
+
+private lemma one_mul'' (x : ⨁ i, A i) : 1 * x = x :=
+suffices mul_hom A 1 = add_monoid_hom.id (⨁ i, A i),
+  from add_monoid_hom.congr_fun this x,
 begin
   apply add_hom_ext, intros i xi,
   unfold has_one.one,
@@ -316,6 +338,64 @@ begin
   unfold has_one.one,
   rw [flip_apply, of_mul_of'],
   exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_one ⟨i, xi⟩),
+end
+
+private lemma mul_assoc''' (a b c : ⨁ i, A i) : a * b * c = a * (b * c) :=
+suffices (mul_hom A).comp_hom.comp (mul_hom A)           -- `λ a b c, a * b * c` as a bundled hom
+   = (add_monoid_hom.comp_hom flip_hom $
+       (mul_hom A).flip.comp_hom.comp (mul_hom A)).flip, -- `λ a b c, a * (b * c)` as a bundled hom
+from sorry,
+-- begin
+--     refine add_monoid_hom.congr_fun _ c,
+--     refine add_monoid_hom.congr_fun _ b,
+--     refine add_monoid_hom.congr_fun _ a,
+--   end,
+begin
+  apply add_hom_ext, intros ai ax, apply add_hom_ext, intros bi bx, apply add_hom_ext, intros ci cx,
+  -- simplify lhs
+  rw [coe_comp, function.comp_app, comp_hom_apply_apply, coe_comp, function.comp_app],
+  -- simplify rhs
+  rw [flip_apply, comp_hom_apply_apply, coe_comp, function.comp_app, flip_hom_apply, coe_comp,
+      function.comp_app, comp_hom_apply_apply, flip_apply, coe_comp, function.comp_app, flip_apply],
+  rw [of_mul_of', of_mul_of', of_mul_of', of_mul_of'],
+  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
+end
+
+private lemma mul_assoc'' (a b c : ⨁ i, A i) : a * b * c = a * (b * c) :=
+suffices (mul_hom A).comp_hom.comp (mul_hom A)           -- `λ a b c, a * b * c` as a bundled hom
+   = (add_monoid_hom.comp_hom flip_hom $
+       (mul_hom A).flip.comp_hom.comp (mul_hom A)).flip, -- `λ a b c, a * (b * c)` as a bundled hom
+from add_monoid_hom.congr_fun (add_monoid_hom.congr_fun (add_monoid_hom.congr_fun this a) b) c,
+begin
+  apply add_hom_ext, intros ai ax, apply add_hom_ext, intros bi bx, apply add_hom_ext, intros ci cx,
+  -- simplify lhs
+  rw [coe_comp, function.comp_app, comp_hom_apply_apply, coe_comp, function.comp_app],
+  -- simplify rhs
+  rw [flip_apply, comp_hom_apply_apply, coe_comp, function.comp_app, flip_hom_apply, coe_comp,
+      function.comp_app, comp_hom_apply_apply, flip_apply, coe_comp, function.comp_app, flip_apply],
+  rw [of_mul_of', of_mul_of', of_mul_of', of_mul_of'],
+  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
+end
+
+private lemma mul_assoc' (a b c : ⨁ i, A i) : a * b * c = a * (b * c) :=
+suffices (mul_hom A).comp_hom.comp (mul_hom A)           -- `λ a b c, a * b * c` as a bundled hom
+   = (add_monoid_hom.comp_hom flip_hom $
+       (mul_hom A).flip.comp_hom.comp (mul_hom A)).flip, -- `λ a b c, a * (b * c)` as a bundled hom
+-- congr_at c $ congr_at b $ congr_at a $
+  -- by { refine add_monoid_hom.congr_fun _ c, },
+-- from add_monoid_hom.congr_fun (add_monoid_hom.congr_fun (add_monoid_hom.congr_fun this c) b) a,
+-- Hmm this works:
+from congr_at c (congr_at b (congr_at a this)),
+-- Ah, ha, it's the order of a, b, c.  See above.
+begin
+  apply add_hom_ext, intros ai ax, apply add_hom_ext, intros bi bx, apply add_hom_ext, intros ci cx,
+  -- simplify lhs
+  rw [coe_comp, function.comp_app, comp_hom_apply_apply, coe_comp, function.comp_app],
+  -- simplify rhs
+  rw [flip_apply, comp_hom_apply_apply, coe_comp, function.comp_app, flip_hom_apply, coe_comp,
+      function.comp_app, comp_hom_apply_apply, flip_apply, coe_comp, function.comp_app, flip_apply],
+  rw [of_mul_of', of_mul_of', of_mul_of', of_mul_of'],
+  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
 end
 
 private lemma mul_assoc (a b c : ⨁ i, A i) : a * b * c = a * (b * c) :=