@gnprice 2021-04-08
This works!
```lean
begin
  ext ai ax bi bx ci cx : 6,
  dsimp,
  rw [of_mul_of', of_mul_of', of_mul_of', of_mul_of'],
  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
end
```

Rather slow, though -- 2.6s.

I'll see if I can find a `dsimp only` that strikes a good balance.