@gnprice 2021-04-08
OK, pushed a new version -- take a look and see what you think.

Here's the new mul_assoc tactic block:
```lean
begin
  ext ai ax bi bx ci cx : 6,
  dsimp only [coe_comp, function.comp_app, comp_hom_apply_apply, flip_apply, flip_hom_apply],
  rw [of_mul_of', of_mul_of', of_mul_of', of_mul_of'],
  exact dfinsupp.single_eq_of_sigma_eq (gmonoid.mul_assoc ⟨ai, ax⟩ ⟨bi, bx⟩ ⟨ci, cx⟩),
end
```

That's a good bit simpler to read than the version with `rw` doing all the work, and it has the nice short `ext` line too. It's a bit slower but not horribly so: 380ms, vs 180ms for the explicit version with `rw`.

I experimented with doing similar things for the other proofs (see next-to-last commit if you want to look), but I think it ends up not being worth it there. Each of those needs only a tiny bit of simplification after the extensionality step -- in particular, there's no existing rewrite steps with `coe_comp` or `function.comp_apply` -- so a `dsimp only` doesn't get to absorb any of that, and ends up not really any simpler than the explicit `apply add_hom_ext, intros …`. And plain `dsimp` is pretty slow, like 1100ms for `mul_comm` vs 85ms for the explicit version.